core_bot/requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
propcache==0.3.1
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
python-dotenv==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.0
yarl==1.18.3
